+++ CHANGELOG.md
@@ -11,6 +11,8 @@ All notable changes to the Audio Player project will be documented in this file.
 
 ### Fixed
 - PHP 8.4 compatibility for nullable parameters
+### Changed
+- Refactor SettingController to use service and mapper
 
 
 ## 3.4.1 - 2023-12-11